Rename heart_beat to heartbeat
