+++ oper8/cmd/__init__.py
@@ -4,6 +4,6 @@ This module holds all of the command classes for oper8's main entrypoint
 
 # Local
 from .base import CmdBase
-from .check_heart_beat import CheckHeartbeatCmd
+from .check_heartbeat import CheckHeartbeatCmd
 from .run_operator_cmd import RunOperatorCmd
 from .setup_vcs_cmd import SetupVCSCmd
